backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
pyjwt>=2.8.0
networkx>=3.2.1
strawberry-graphql[fastapi]>=0.213.0
redis>=5.0.1
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-pptx>=0.6.23
